src/pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
replit = "^3.2.4"
Flask = "^2.2.0"
urllib3 = "^1.26.12"
cryptography = "^38.0.2"
pwinput = "^1.0.2"
bcrypt = "^4.0.1"
pycryptodome = "^3.15.0"
thread6 = "^0.2.0"
thread-py = "^1.0.1"
thread_utils = "^1.0.1"
parse = "^1.19.0"
psutil = "^5.9.4"
pyOpenSSL = "^22.1.0"

[tool.poetry.dev-dependencies]
debugpy = "^1.6.2"
replit-python-lsp-server = {extras = ["yapf", "rope", "pyflakes"], version = "^1.5.9"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"